ARALILA/backend/requirements.txt
aioredis>=2.0.0
asgiref
channels
channels_redis
daphne
Django
dj-database-url==2.1.0
django-extensions
django-cors-headers
djangorestframework
djangorestframework-simplejwt
openai>=1.0.0
PyJWT
pytz
psycopg2-binary
python-dotenv
sqlparse
whitenoise
redis>=5.0.0
websockets